api/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
